learn C programing

+++ 0x00-hello_world/6-size.c
@@ -8,7 +8,7 @@
 int main(void)
 {
 	printf("Size of Char Data Types in C = %zu bytes \n", sizeof(char));
-	printf("Size of Int Data Types in C = %zu bytes \n", sizeof(short int));
+	printf("Size of Int Data Types in C = %zu bytes \n", sizeof(int));
 	printf("Size of Long Int Data Types in C = %zu bytes \n", sizeof(long int));
 	printf("Size of Long Long Int Data Types in C = %zu bytes \n", sizeof(long long int));
 	printf("Size of Float Data Types in C = %zu bytes \n", sizeof(float));
